moves useAddresses hook to own file
